requirements.txt
boto3==1.3.1
Flask==0.10.1
Jinja2==2.8
Werkzeug==0.11.3
pymongo==3.8.0
uwsgi
python-dotenv
